% Auto-generated: do not edit by hand
\name{button}

\alias{button}

\title{button component}

\description{

}

\usage{
button(children=NULL, title=NULL, className=NULL, id=NULL,
onClick=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Children}

\item{title}{Character. The title of the Button}

\item{className}{Character. The CSS class(es)}

\item{id}{Character. The ID}

\item{onClick}{The onClick function}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

