% Auto-generated: do not edit by hand
\name{controls}

\alias{controls}

\title{controls component}

\description{

}

\usage{
controls(id=NULL, zoom=NULL)
}

\arguments{
\item{id}{Character. The id of the SVG}

\item{zoom}{Named list. The zoom object}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

