% Auto-generated: do not edit by hand
\name{fetchData}

\alias{fetchData}

\title{fetchData component}

\description{
FetchData is an FetchData example component
}

\usage{
fetchData(id=NULL, filterOut=NULL)
}

\arguments{
\item{id}{Character. The id displayed in the input.}

\item{filterOut}{Character. The color to filter out.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

