% Auto-generated: do not edit by hand
\name{heatmap}

\alias{heatmap}

\title{heatmap component}

\description{
Heatmap component. It takes an warehouse layout (svg)  and overlays the data supplied
}

\usage{
heatmap(id=NULL, width=NULL, svg=NULL, data=NULL)
}

\arguments{
\item{id}{Character. The id of the SVG}

\item{width}{Numeric. The width of the SVG}

\item{svg}{Character. The SVG image.}

\item{data}{Named list. The data object that generates the SVG}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

