% Auto-generated: do not edit by hand
\name{mapKey}

\alias{mapKey}

\title{mapKey component}

\description{
Map Key Components
}

\usage{
mapKey(colorMap=NULL)
}

\arguments{
\item{colorMap}{Named list. The colorMap object}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

