% Auto-generated: do not edit by hand
\name{svgControls}

\alias{svgControls}

\title{svgControls component}

\description{
Zoom and Pan Controls Component
}

\usage{
svgControls(id=NULL, zoom=NULL)
}

\arguments{
\item{id}{Character. The id of the SVG}

\item{zoom}{Named list. The zoom object}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

